move the "response" files to "service" package. remove some useless code & files
