--- android/app/src/main/java/com/example/medimind/OnEditMedicineRequestedListener.kt
@@ -1,5 +1,0 @@
-package com.example.medimind
-
-interface OnEditMedicineRequestedListener {
-    fun onEditMedicineRequested(medicineName: String)
-}
