--- android/app/src/main/java/com/example/medimind/OnMedicineSelectedListener.kt
@@ -1,5 +1,0 @@
-package com.example.medimind
-
-interface OnMedicineSelectedListener {
-    fun onMedicineSelected(medicineName: String)
-}
