+++ android/app/src/main/java/com/example/medimind/service/EditMedResponse.kt
@@ -1,4 +1,4 @@
-package com.example.medimind.data
+package com.example.medimind.service
 
 data class EditMedResponse(
     val frequency: Int,
