+++ android/app/src/main/java/com/example/medimind/service/IntakeHistoryResponse.kt
@@ -1,4 +1,4 @@
-package com.example.medimind.data
+package com.example.medimind.service
 
 data class IntakeHistoryResponse(
     val medicationName: String,
