+++ android/app/src/main/java/com/example/medimind/service/MedicationResponse.kt
@@ -1,6 +1,4 @@
-package com.example.medimind.data
-
-import com.google.gson.annotations.SerializedName
+package com.example.medimind.service
 
 data class MedicationResponse(
     val id: String,
