+++ android/app/src/main/res/layout/fragment_edit_medicine_details.xml
@@ -33,6 +33,7 @@
             android:layout_width="60dp"
             android:layout_height="wrap_content"
             android:inputType="number"
+            android:maxLength="2"
             android:hint="3"
             android:layout_marginStart="8dp"/>
